refactor(aaa): restructure into avatar mod
