+++ pallets/ajuna-awesome-avatars/src/types/avatar/force.rs
@@ -18,40 +18,6 @@ use crate::*;
 use codec::alloc::string::ToString;
 use sp_std::{fmt, prelude::*};
 
-#[derive(Encode, Decode, Debug, Default, PartialEq)]
-pub struct AvatarCodec {
-	pub season_id: SeasonId,
-	pub dna: Dna,
-	pub soul_points: SoulCount,
-	pub rarity: BoundedVec<u8, ConstU32<20>>,
-	pub force: BoundedVec<u8, ConstU32<20>>,
-}
-
-impl From<Avatar> for AvatarCodec {
-	fn from(avatar: Avatar) -> Self {
-		let rarity_tier = RarityTier::try_from(avatar.min_tier()).unwrap_or_default();
-		let last_variation = Force::try_from(avatar.last_variation()).unwrap_or_default();
-
-		Self {
-			season_id: avatar.season_id,
-			dna: avatar.dna.clone(),
-			soul_points: avatar.souls,
-			rarity: rarity_tier.into(),
-			force: last_variation.into(),
-		}
-	}
-}
-
-impl From<AvatarCodec> for Avatar {
-	fn from(avatar_codec: AvatarCodec) -> Self {
-		Self {
-			season_id: avatar_codec.season_id,
-			dna: avatar_codec.dna,
-			souls: avatar_codec.soul_points,
-		}
-	}
-}
-
 pub enum Force {
 	Kinetic = 0,
 	Dream = 1,
@@ -64,6 +30,19 @@ pub enum Force {
 impl Default for Force {
 	fn default() -> Self {
 		Force::Kinetic
+	}
+}
+
+impl fmt::Display for Force {
+	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+		match self {
+			Force::Kinetic => write!(f, "Kinetic"),
+			Force::Dream => write!(f, "Dream"),
+			Force::Solar => write!(f, "Solar"),
+			Force::Thermal => write!(f, "Thermal"),
+			Force::Astral => write!(f, "Astral"),
+			Force::Empathy => write!(f, "Empathy"),
+		}
 	}
 }
 
@@ -82,21 +61,9 @@ impl TryFrom<u8> for Force {
 		}
 	}
 }
+
 impl From<Force> for BoundedVec<u8, ConstU32<20>> {
 	fn from(x: Force) -> Self {
 		x.to_string().as_bytes().to_owned().try_into().unwrap_or_default()
-	}
-}
-
-impl fmt::Display for Force {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		match self {
-			Force::Kinetic => write!(f, "Kinetic"),
-			Force::Dream => write!(f, "Dream"),
-			Force::Solar => write!(f, "Solar"),
-			Force::Thermal => write!(f, "Thermal"),
-			Force::Astral => write!(f, "Astral"),
-			Force::Empathy => write!(f, "Empathy"),
-		}
 	}
 }
